### 2018-08-28 #917 @whedon
[PRE REVIEW]: hIPPYlib: An Extensible Software Framework for Large-Scale Inverse Problems
**Submitting author:** @uvilla (<a href="http://orcid.org/0000-0002-5142-2559">Umberto Villa</a>)
**Repository:** <a href="https://github.com/hippylib/hippylib" target ="_blank">https://github.com/hippylib/hippylib</a>
**Version:** 2.1.0
