### 2018-08-28 d3cc396
Update paper.md
