### 2018-08-28 #917 @uvilla
Suggested reviewers

- @IgorBaratta: `hIPPYlib` relies on `FEniCS` for the discretization of the PDE, and for generating weak forms for the adjoint and incrementals problem. Since @IgorBaratta is a `FEniCS` developer he might be interested in reviewing `hIPPYlib`.

- @lheagy: She is working on inverse problem with geophysical applications
