### 2018-08-30 #917 @lheagy
:wave: Hi @IgorBaratta, would you be willing to review this submission for JOSS?
