### 2018-08-30 #917 @IgorBaratta
> Hi @IgorBaratta, would you be willing to review this submission for JOSS?

Hi @lheagy , I'd be glad to review this submission.
