### 2018-09-12 e19a2d1
Update paper.bib
[skip ci]
