+++ paper/paper.bib
@@ -6,7 +6,7 @@
 }
 
 @book{LangtangenLogg17,
- title = {Solving PDEs in Python},
+ title = {Solving {PDE}s in {P}ython},
  author = {Hans Petter Langtangen and Anders Logg},
  year = {2017},
  publisher = {Springer},
@@ -19,7 +19,9 @@
   Author                   = {Logg, Anders and Mardal, Kent-Andre and Wells, Garth},
   Publisher                = {Springer Science \& Business Media},
   Year                     = {2012},
-  Volume                   = {84}
+  Volume                   = {84},
+  doi = {10.1007/978-3-642-23099-8},
+  isbn = {978-3-642-23098-1}
 }
 
 @Article{LoggWells10,
