### 2018-09-21 5b817ab
Update README.md
Add JOSS status
