### 2018-09-26 #940 @boehmc
Hey Umberto (and co),
HippyLib is a really cool package and nicely put together! Being able to run everything in a docker image / jupyter nb is just awesome. Thanks a lot for all the effort that went into it!!

Please allow me a few more days to play around with it. I will finish everything by the end of the week.

In the mean time I have a few minor things:

- Could you add a doi for the DOLFIN paper (10.1145/1731022.1731030)?

- The doi's of the bibtex book items do not appear in the pdf output. I think this could be an issue with the bib style. But I don't know whether it makes sense to include them...

- Do you know the package nbval to test jupyter notebooks (https://nbval.readthedocs.io/en/latest/)? If I am not mistaken you have set up a CI on travis that runs all the notebooks, but it doesn't check the output. Using nbval is quite handy for this and maybe something you may consider in the future.

- I found a few typos in the jupyter notebook (nothing important). I could either turn this into a pull request or just send you the git diff. Please let me know what you prefer.

 **Paper**
- I think the term "PDE-constrained" might be more common than "PDE-based"

- The use of the word *automatic* in the comment on gradient and Hessian evaluations might be a bit misleading / contradicting when looking at the third and fifth paragraph. If I understood correctly, the user needs to provide the weak form of all forward / adjoint equations and FEniCS will *automatically* set up and solve the resulting PDEs, whereas in the future you might even skip the necessity of specifying the weak form of the adjoint equations, because AD techniques will *automatically* generate it for the user. Maybe you could slightly reformulate this sentence.

- Since you are using Steihaug-CG, I was wondering if there is actually a need for Armijo back-tracking? In my experience, Steihaug-CG is often combined with a trust-region globalization.

- *"low rank factorisation of the Hessian misfit"* sounds a bit odd to me. Maybe *Hessian of the misfit term* or just *Hessian*?

- I don't know if there are any space limitations, but I would consider adding references to part 1 and 2 of the SIAM papers of your co-authors on the "Computatial framework for infinite-dimensional Bayesian inverse problems. For people not familiar with this field, this would be an excellent starting point to get a better idea of all the math that is going on in the background.
