### 2018-09-26 #940 @uvilla
Hi @boehmc ,

Thank you very much for taking the time to look at the code and for your *spot on* comments.
I am happy that you found `hIPPYlib` *really cool* and the docker image easy to use.

- I agree with all your comments on the paper, and I will revise it accordingly.
In particular, thank you for catching the misleading statement about *automatic* differentiation. hIPPYlib uses FEniCS's *symbolic* differentiation of weak forms, but we do not use any of the automatic differentiation tools in `dolfin-adjoint`. I will fix that.
Trust-region globalization was actually implemented recently in `hIPPYlib`. I will mention that in the revised paper. Overall, in my experience, the two types of globalization perform quite similarly.
I will also add the references you mentioned and change the odd wordings .Missing DOIs should get fixed once I recompile the paper.

- Thank you also for suggesting `nbval` to check the notebook output for correctness. I'll look into it.

- Finally, thank you for finding some typos in the notebooks and fixing them. If it is not too much work on your hand, a pull request is very appreciated ;)

No rush to complete the review. Please take your time experimenting with the code.

Cheers,

Umberto
