### 2018-09-27 1b862f0
[joss-review][skip ci]
Address @boehmc 's comments regarding the JOSS manuscript.

Specifically:

1. Use "PDE constrained" instead of "PDE based"
2. Clarify how gradient and Hessian actions are computed in hIPPYlib using FEniCS *symbolic* differentiation of weak form. This should clarify the confusing and misleading statements regarding automatic differentiation
3. Mention that hIPPYlib also provides trust region as globalization technique
4. Replace odd language *Hessian misfit* with *Hessian of log likelihood*
5. Add a references to *Computational framework for infinite-dimensional Bayesian 
[…]
