+++ paper/paper.bib
@@ -33,3 +33,25 @@
   Pages                    = {Article 20},
   Volume                   = {37}
 }
+
+@Article{Bui-ThanhGhattasMartinEtAl13,
+  Title                    = {A Computational Framework for Infinite-Dimensional {B}ayesian Inverse Problems {P}art {I}: {T}he Linearized Case, with Application to Global Seismic Inversion},
+  Author                   = {Bui-Thanh, Tan and Ghattas, Omar and Martin, James and Stadler, Georg},
+  Journal                  = {SIAM Journal on Scientific Computing},
+  Year                     = {2013},
+  Number                   = {6},
+  Pages                    = {A2494-A2523},
+  Volume                   = {35},
+  doi                      = {10.1137/12089586X}
+}
+
+@Article{PetraMartinStadlerEtAl14,
+  Title                    = {A computational framework for infinite-dimensional {B}ayesian inverse problems: {P}art {II}. {S}tochastic {N}ewton {MCMC} with application to ice sheet inverse problems},
+  Author                   = {Noemi Petra and James Martin and Georg Stadler and Omar Ghattas},
+  Journal                  = {SIAM Journal on Scientific Computing},
+  Year                     = {2014},
+  Number                   = {4},
+  Pages                    = {A1525--A1555},
+  Volume                   = {36},
+  doi = {10.1137/130934805}
+}
