+++ README.md
@@ -18,8 +18,9 @@ $$/   $$/ $$$$$$/ $$/       $$/           $$/     $$/ $$/ $$$$$$$/
 
                   https://hippylib.github.io
 
-`hIPPYlib` implements state-of-the-art scalable algorithms for PDE-based
-deterministic and Bayesian inverse problems. It builds on [FEniCS](http://fenicsproject.org/)
+`hIPPYlib` implements state-of-the-art scalable algorithms for
+deterministic and Bayesian inverse problems governed by partial differential equations (PDE).
+It builds on [FEniCS](http://fenicsproject.org/)
 (a parallel finite element element library) for the discretization of the PDE
 and on [PETSc](http://www.mcs.anl.gov/petsc/) for scalable and efficient linear
 algebra operations and solvers.
@@ -32,42 +33,63 @@ features are the interactive tutorials in the `tutorial folder`.
 
 Conceptually, `hIPPYlib` can be viewed as a toolbox that provides the
 building blocks for experimenting new ideas and developing scalable
-algorithms for PDE-based deterministic and Bayesian inverse problems.
+algorithms for PDE constrained deterministic and Bayesian inverse problems.
 
 In `hIPPYlib` the user can express the forward PDE and the likelihood in
 weak form using the friendly, compact, near-mathematical notation of
 `FEniCS`, which will then automatically generate efficient code for the
 discretization.  Linear and nonlinear, and stationary and
-time-dependent PDEs are supported in `hIPPYlib`. Currently, gradient and
-Hessian information needs to be provided by the user; our future plan
-includes automatic generation of this information by using `FEniCS`
-automatic differentiation capabilities for users who do not wish to
-derive the relevant weak forms.
+time-dependent PDEs are supported in `hIPPYlib`.
+For stationary problems, gradient and Hessian information can be
+automatically generated by hIPPYlib using FEniCS symbolic differentiation
+of the relevant weak forms. For time-dependent problems, instead, symbolic
+differentiation can only be used for the spatial terms, and the contribution
+to gradients and Hessians arising from the time dynamics needs to be provided
+by the user.
 
 Noise and prior covariance operators are modeled as inverses of
 elliptic differential operators allowing us to build on existing fast
 multigrid solvers for elliptic operators without explicitly
 constructing the dense covariance operator.
 
+The key property of the algorithms underlying hIPPYlib is that solution
+of the deterministic and linearized Bayesian inverse problem is computed
+at a cost, measured in forward PDE solves, that is independent of the
+parameter dimension.
+
 `hIPPYlib` provides a robust implementation of the inexact
 Newton-conjugate gradient algorithm to compute the maximum a posterior
-(MAP) point. The reduced gradient and Hessian actions are
-automatically computed via their weak form specification in `FEniCS` by
+(MAP) point. The gradient and Hessian actions are
+computed via their weak form specification in `FEniCS` by
 constraining the state and adjoint variables to satisfy the forward
 and adjoint problem. The Newton system is solved inexactly by early
 termination of CG iterations via Eisenstat-Walker (to prevent
 oversolving) and Steihaug (to avoid negative curvature)
-criteria. Globalization is achieved with an Armijo back-tracking line
-search.
+criteria. Two globalization techniques are available to the user:
+Armijo back-tracking line search and trust region.
 
-`hIPPYlib` offers different scalable methods to sample from the prior
-distribution: Krylov methods to approximate the action of a matrix
-square root on a vector, the conjugate gradient sampler, and finally
-samplers that exploit the finite element assembly procedure of the
-inverse covariance operator to obtain a symmetric decomposition. To
-sample from a local Gaussian approximation (such as at the MAP point)
-`hIPPYlib` exploits the low rank factorization of the Hessian misfit to
[…]
