### 2018-09-27 0a7ed1a
[joss-review] [skip ci]
Add missing DOI to reference @LoggWells10
