+++ paper/paper.bib
@@ -32,6 +32,7 @@
   Number                   = {2},
   Pages                    = {Article 20},
   Volume                   = {37}
+  doi                      = {10.1145/1731022.1731030}
 }
 
 @Article{Bui-ThanhGhattasMartinEtAl13,
