### 2018-09-27 67dada6
[joss-review] [skip ci]
Fix failed pdf compilation
