+++ paper/paper.bib
@@ -31,7 +31,7 @@
   Year                     = {2010},
   Number                   = {2},
   Pages                    = {Article 20},
-  Volume                   = {37}
+  Volume                   = {37},
   doi                      = {10.1145/1731022.1731030}
 }
 
