### 2018-09-27 #940 @uvilla
@boehmc ,

Thank you again for your constructive comments regarding the paper. Commit [1b862f03d633a0da8eb1ab62d4d3b62d4da22563](https://github.com/hippylib/hippylib/commit/1b862f03d633a0da8eb1ab62d4d3b62d4da22563) addresses them.

Here, the revised [manuscript](https://github.com/openjournals/joss-papers/blob/joss.00940/joss.00940/10.21105.joss.00940.pdf).

Looking forward for the rest of your review.
