### 2018-09-28 7e3048b
[joss-review] [skip ci]
Minor edits
