+++ README.md
@@ -19,7 +19,7 @@ $$/   $$/ $$$$$$/ $$/       $$/           $$/     $$/ $$/ $$$$$$$/
                   https://hippylib.github.io
 
 `hIPPYlib` implements state-of-the-art scalable algorithms for
-deterministic and Bayesian inverse problems governed by partial differential equations (PDE).
+deterministic and Bayesian inverse problems governed by partial differential equations (PDEs).
 It builds on [FEniCS](http://fenicsproject.org/)
 (a parallel finite element element library) for the discretization of the PDE
 and on [PETSc](http://www.mcs.anl.gov/petsc/) for scalable and efficient linear
@@ -33,7 +33,7 @@ features are the interactive tutorials in the `tutorial folder`.
 
 Conceptually, `hIPPYlib` can be viewed as a toolbox that provides the
 building blocks for experimenting new ideas and developing scalable
-algorithms for PDE constrained deterministic and Bayesian inverse problems.
+algorithms for PDE-constrained deterministic and Bayesian inverse problems.
 
 In `hIPPYlib` the user can express the forward PDE and the likelihood in
 weak form using the friendly, compact, near-mathematical notation of
@@ -41,7 +41,7 @@ weak form using the friendly, compact, near-mathematical notation of
 discretization.  Linear and nonlinear, and stationary and
 time-dependent PDEs are supported in `hIPPYlib`.
 For stationary problems, gradient and Hessian information can be
-automatically generated by hIPPYlib using FEniCS symbolic differentiation
+automatically generated by `hIPPYlib` using `FEniCS` symbolic differentiation
 of the relevant weak forms. For time-dependent problems, instead, symbolic
 differentiation can only be used for the spatial terms, and the contribution
 to gradients and Hessians arising from the time dynamics needs to be provided
@@ -52,8 +52,8 @@ elliptic differential operators allowing us to build on existing fast
 multigrid solvers for elliptic operators without explicitly
 constructing the dense covariance operator.
 
-The key property of the algorithms underlying hIPPYlib is that solution
-of the deterministic and linearized Bayesian inverse problem is computed
+The key property of the algorithms underlying `hIPPYlib` is that solution
+of the deterministic and Bayesian inverse problem is computed
 at a cost, measured in forward PDE solves, that is independent of the
 parameter dimension.
 
@@ -78,13 +78,12 @@ methods. `hIPPYlib` makes the construction of the posterior covariance
 tractable by invoking a low-rank approximation of the Hessian of the
 log likelihood.
 
-`hIPPYlib` offers different scalable methods for sample generation.
+`hIPPYlib` also offers scalable methods for sample generation.
 To sample large scale spacially correlated Gaussian random fields from the prior
 distribution, `hIPPYlib`  implements a new method that strongly relies on the
-structure of the covariance operator (defined as the inverse of a differential operator)
-by exploiting the assembly procedure of finite element matrices to construct a sparse
-rectagular decomposition of the precision operator.
-To sample from a local Gaussian approximation (such as at the MAP point)
+structure of the covariance operator defined as the inverse of a differential operator:
+by exploiting the assembly procedure of finite element matrices `hIPPYlib` constructs a sparse Cholesky-like rectagular decomposition of the precision operator.
+To sample from a local Gaussian approximation to the posterior (such as at the MAP point)
 `hIPPYlib` exploits the low rank factorization of the Hessian of the
 log likelihood to correct samples from the prior distribution.
 Finally, to explore the posterior distribution, `hIPPYlib` implements
