### 2018-10-02 #940 @boehmc
@uvilla,

I've just sent you a pull requests with a few tiny typos I noticed while reviewing the paper (only one of them is actually math related)

Additionally, in tutorial 4, you may want to switch the notation of the domain from D to Omega and I think the forward operator is sometimes denoted by F and sometimes by B.

Thanks for updating the paper. Everything looks fine to me and it is very convenient to use with the docker image. This is a really nice package and I have a few applications in mind, where I might use hippylib for my own research. Thanks a lot for making it available!
