### 2018-10-03 #940 @IgorBaratta
I had known hIPPYlib for some time and recently played with the tutorials. The tutorials are informative and indeed an excellent place to start to learn about the package.

Great work @uvilla  and colleagues!

----
Some small _suggestions_ to the software paper
1st paragraph:
-  Replace the first parenthesis with a comma to improve readability.
- Include a reference to the PETSc library, to be consistent with the other dependencies that have been credited.

8th paragraph:
- correct a small typo: rectagular – rectangular
-----

I could not find any community guidelines on the Github repository nor the Github page (https://hippylib.github.io/). If this information is not available, I would recommend creating a simple guideline on how to contribute to the package, report issues, get support, etc..
