### 2018-10-03 be26f75
[joss-review] [skip ci]
Address @@IgorBaratta 's comments:
- Fix punctuation in the first sentence
- Add reference to PETSc
- Correct mispelled word.
