+++ paper/paper.bib
@@ -7,7 +7,7 @@
 
 @book{LangtangenLogg17,
  title = {Solving {PDE}s in {P}ython},
- author = {Hans Petter Langtangen and Anders Logg},
+ author = {Langtangen, Hans Petter and Logg, Anders},
  year = {2017},
  publisher = {Springer},
  doi = {10.1007/978-3-319-52462-7},
@@ -26,7 +26,7 @@
 
 @Article{LoggWells10,
   Title                    = {{DOLFIN}: Automated Finite Element Computing},
-  Author                   = {Anders Logg and Garth N. Wells},
+  Author                   = {Logg, Anders and Wells, Garth N. },
   Journal                  = {ACM Transactions on Mathematical Software},
   Year                     = {2010},
   Number                   = {2},
@@ -48,11 +48,23 @@
 
 @Article{PetraMartinStadlerEtAl14,
   Title                    = {A computational framework for infinite-dimensional {B}ayesian inverse problems: {P}art {II}. {S}tochastic {N}ewton {MCMC} with application to ice sheet inverse problems},
-  Author                   = {Noemi Petra and James Martin and Georg Stadler and Omar Ghattas},
+  Author                   = {Petra, Noemi and Martin, James and Stadler, Georg and Ghattas, Omar},
   Journal                  = {SIAM Journal on Scientific Computing},
   Year                     = {2014},
   Number                   = {4},
   Pages                    = {A1525--A1555},
   Volume                   = {36},
   doi = {10.1137/130934805}
+}
+
+@Misc{petsc-web-page,
+  author = {Balay, Satish  and Abhyankar, Shrirang and Adams, Mark~F. and Brown, Jed and Brune, Peter 
+            and Buschelman, Kris and Dalcin,Lisandro and Dener, Alp and Eijkhout, Victorand 
+            and Gropp, William~D. and Kaushik, Dinesh and Knepley, Matthew~G. and May, Dave~A. 
+            and McInnes, Lois Curfman and Mills, Richard Tran and Munson, Todd and Rupp, Karl 
+            and Sanan, Patrick and Smith, Barry~F. and Zampini, Stefano and Zhang, Hong},
+   title =  {{PETS}c {W}eb page},
+   url =    {http://www.mcs.anl.gov/petsc},
+   howpublished = {\url{http://www.mcs.anl.gov/petsc}},
+   year = {2018}
 }
