### 2018-10-03 e714ca1
[joss-review] [skip ci]
Fix spacing
