### 2018-10-10 #940 @uvilla
@IgorBaratta thank you for your review and kind comments.

> Some small suggestions to the software paper

I have made the suggested changes see commit [be26f75b66399c9bb6a00238cc3fe71e0a80326e](https://github.com/hippylib/hippylib/commit/be26f75b66399c9bb6a00238cc3fe71e0a80326e).

> I could not find any community guidelines on the Github repository nor the Github page.

You are correct there is no community guidelines yet. I am drafting a CONTRIBUTING.md that will be added to the source code and published---together with the sphinx documentation--- on readthedoc.
I'll let you know when this task is completed.

Thank you again,

Umberto
