### 2018-10-10 372c53b
[joss-review] [skip ci]
First draft of contributing guidelines
