+++ CONTRIBUTING.md
@@ -1,0 +1,262 @@
+[![Build Status](https://travis-ci.org/hippylib/hippylib.svg?branch=master)](https://travis-ci.org/hippylib/hippylib)
+[![status](http://joss.theoj.org/papers/053e0d08a5e9755e7b78898cff6f6208/status.svg)](http://joss.theoj.org/papers/053e0d08a5e9755e7b78898cff6f6208) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.596931.svg)](https://doi.org/10.5281/zenodo.596931)
+
+                  Inverse Problem PYthon library
+
+```
+ __        ______  _______   _______   __      __  __  __  __       
+/  |      /      |/       \ /       \ /  \    /  |/  |/  |/  |      
+$$ |____  $$$$$$/ $$$$$$$  |$$$$$$$  |$$  \  /$$/ $$ |$$/ $$ |____  
+$$      \   $$ |  $$ |__$$ |$$ |__$$ | $$  \/$$/  $$ |/  |$$      \ 
+$$$$$$$  |  $$ |  $$    $$/ $$    $$/   $$  $$/   $$ |$$ |$$$$$$$  |
+$$ |  $$ |  $$ |  $$$$$$$/  $$$$$$$/     $$$$/    $$ |$$ |$$ |  $$ |
+$$ |  $$ | _$$ |_ $$ |      $$ |          $$ |    $$ |$$ |$$ |__$$ |
+$$ |  $$ |/ $$   |$$ |      $$ |          $$ |    $$ |$$ |$$    $$/ 
+$$/   $$/ $$$$$$/ $$/       $$/           $$/     $$/ $$/ $$$$$$$/  
+```                                                                    
+                                                                    
+
+                  https://hippylib.github.io
+                  
+# How to Contribute
+
+The `hIPPYlib` team welcomes contributions at all levels: bugfixes, code
+improvements, new capabilities, improved documentation, 
+or new examples/tutorials.
+
+Use a pull request (PR) toward the `hippylib:master` branch to propose your
+contribution. If you are planning significant code changes, or have any
+questions, you should also open an [issue](https://github.com/hippylib/hippylib/issues)
+before issuing a PR. 
+
+See the [Quick Summary](#quick-summary) section for the main highlights of our
+GitHub workflow. For more details, consult the following sections and refer
+back to them before issuing pull requests:
+
+- [GitHub Workflow](#github-workflow)
+  - [hIPPYlib Organization](#hippylib-organization)
+  - [New Feature Development](#new-feature-development)
+  - [Developer Guidelines](#developer-guidelines)
+  - [Pull Requests](#pull-requests)
+  - [Pull Request Checklist](#pull-request-checklist)
+- [Automated Testing](#automated-testing)
+- [Contact Information](#contact-information)
+
+Contributing to hIPPYlib requires knowledge of Git and, likely, inverse problems.
+If you are new to Git, see the [GitHub learning
+resources](https://help.github.com/articles/git-and-github-learning-resources/).
+To learn more about inverse problems, see our [tutorial page](http://hippylib.github.io/tutorial).
+
+*By submitting a pull request, you are affirming the [Developer's Certificate of
+Origin](#developers-certificate-of-origin-11) at the end of this file.*
+
+
+## Quick Summary
+
+- We encourage you to [join the hIPPYlib organization](#hippylib-organization) and create
+  development branches off `hippylib:master`.
+- Please follow the [developer guidelines](#developer-guidelines), in particular
+  with regards to documentation and code styling.
+- Pull requests  should be issued toward `hippylib:master`. Make sure
+  to check the items off the [Pull Request Checklist](#pull-request-checklist).
[…]
