### 2018-10-10 bd52ef7
Update CONTRIBUTING.md
[joss-review] [skip ci]
