### 2018-10-11 #940 @uvilla
@IgorBaratta,

I added contributing guidelines both in the GitHub repository (see CONTRIBUTING.md) and on readthedocs (see https://hippylib.readthedocs.io/en/latest/contributing.html).

Please let me know if there are other issues you’d like to be addressed before accepting hIPPYlib.
