### 2018-10-12 db71e8d
Added forward_uq doc
