+++ doc/hippylib.forward_uq.rst
@@ -1,0 +1,78 @@
+hippylib.forward\_uq package
+============================
+
+Submodules
+----------
+
+hippylib.forward\_uq.qoi module
+-------------------------------
+
+.. automodule:: hippylib.forward_uq.qoi
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+hippylib.forward\_uq.reduced\_qoi module
+----------------------------------------
+
+.. automodule:: hippylib.forward_uq.reduced_qoi
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+hippylib.forward\_uq.taylor\_approx\_qoi module
+-----------------------------------------------
+
+.. automodule:: hippylib.forward_uq.taylor_approx_qoi
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+hippylib.forward\_uq.varianceReductionMC module
+-----------------------------------------------
+
+.. automodule:: hippylib.forward_uq.varianceReductionMC
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+hippylib.forward\_uq.rare_event module
+-----------------------------------------------
+
+.. automodule:: hippylib.forward_uq.rare_events.bayesISLowRank
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: hippylib.forward_uq.rare_events.bimcOptProblem
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: hippylib.forward_uq.rare_events.linearApprox
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: hippylib.forward_uq.rare_events.optLikeVar
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: hippylib.forward_uq.rare_events.rareEventMC
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: hippylib.forward_uq.rare_events.utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: hippylib.forward_uq
+    :members:
+    :undoc-members:
+    :show-inheritance:
