### 2018-10-12 b164cbf
Added CI tests
