### 2018-10-12 2b72ca1
Edited CHANGELOG
