### 2018-10-12 4b0bfbe
Fix broken file names in .travis.yml
Update .travis.yml

Small edits to test to be run

Yet another .travis.yml fix

Update travis.yml
