### 2018-10-13 60cbfe3
[joss-review] [skip ci]
Fix spelling error
