+++ paper/paper.md
@@ -84,7 +84,7 @@ tractable by invoking a low-rank approximation of the Hessian of the
 log likelihood.
 
 `hIPPYlib` also offers scalable methods for sample generation.
-To sample large scale spacially correlated Gaussian random fields from the prior
+To sample large scale spatially correlated Gaussian random fields from the prior
 distribution, `hIPPYlib`  implements a new method that strongly relies on the
 structure of the covariance operator defined as the inverse of a differential operator:
 by exploiting the assembly procedure of finite element matrices `hIPPYlib` constructs a sparse Cholesky-like rectangular decomposition of the precision operator.
