### 2018-10-13 943d0cc
Update contributing.rst
[joss-review] [skip ci]
