### 2018-10-13 63ab8ca
Update CONTRIBUTING.md
[joss-review] [skip ci]

Fix a conversion to .rst problem
