### 2018-10-13 c6ac958
Update CONTRIBUTING.md
add docs status
