### 2018-10-18 #940 @uvilla
@IgorBaratta ,

Did you have a change to review our contributing guidelines?
Is there any other issue that you'd like to be address?
Thank you again for your time in reviewing hIPPYlib and for your constructive comments.
