### 2018-10-18 bb89b0b
Update INSTALL.md
[skip ci]

Update instructions for Anaconda
