### 2018-10-18 2d8cb37
Update .travis.yml
Run applications script with `mpi`
