### 2018-10-19 7917340
Added GaussianRealPrior
