+++ hippylib/modeling/prior.py
@@ -15,6 +15,7 @@ from __future__ import absolute_import, division, print_function
 
 import dolfin as dl
 import numpy as np
+import scipy.linalg as scila
 import math
 
 from ..algorithms.linalg import MatMatMult, get_diagonal, amg_method, estimate_diagonal_inv2, Solver2Operator, Operator2Solver
@@ -605,3 +606,129 @@ class MollifiedBiLaplacianPrior(_Prior):
         
         if add_mean:
             s.axpy(1., self.mean)
+
+
+class GaussianRealPrior(_Prior):
+    """
+    This class implements a finite-dimensional Gaussian prior,
+    :math:`\\mathcal{N}(\\boldsymbol{m}, \\boldsymbol{C})`, where
+    :math:`\\boldsymbol{m}` is the mean of the Gaussian distribution, and
+    :math:`\\boldsymbol{C}` is its covariance. The underlying finite element
+    space is assumed to be the "R" space.
+    """
+
+    def __init__(self, Vh, covariance, mean=None):
+        """
+        Constructor
+
+        Inputs:
+        - :code:`Vh`:             Finite element space on which the prior is
+                                  defined. Must be the Real space with one global 
+                                  degree of freedom
+        - :code:`covariance`:     The covariance of the prior. Must be a
+                                  :code:`numpy.ndarray` of appropriate size
+        - :code:`mean`(optional): Mean of the prior distribution. Must be of
+                                  type `dolfin.Vector()`
+        """
+
+        self.Vh = Vh
+
+        if Vh.dim() != covariance.shape[0] or Vh.dim() != covariance.shape[1]:
+            raise ValueError("Covariance incompatible with Finite Element space")
+
+        self.covariance = covariance
+        
+        #np.linalg.cholesky automatically provides error checking, so use those
+        self.chol = np.linalg.cholesky(self.covariance)
+
+        #log of the determinant of the covariance
+        self.log_det_cov = 2. * np.sum(np.log(np.diag(self.chol)))
+
+        self.chol_inv = scila.solve_triangular(
+                                        self.chol,
+                                        np.identity(Vh.dim()),
+                                        lower=True)
+
+        self.precision = np.dot(self.chol_inv.T, self.chol_inv)
+
+        trial = dl.TrialFunction(Vh)
+        test  = dl.TestFunction(Vh)
+        
+        #mass matrix
+        self.M = dl.assemble(dl.inner(trial, test) * dl.dx)
+        self.Msolver = Operator2Solver(self.M)
+
+        if mean:
+            self.mean = mean
+        else:
+            tmp = dl.Vector()
+            self.M.init_vector(tmp, 0)
+            tmp.zero()
+            self.mean = tmp
+
+        if Vh.dim() == 1:
+
+            trial = [trial]
+            test = [test]
+
+        #variational for the regularization operator, or the precision matrix
+        var_form_R = sum([self.precision[i,j] * trial[i] * test[j]
+                          for i in range(Vh.dim()) for j in range(Vh.dim())])
+
+        #variational for the inverse regularization operator, or the covariance
+        #matrix
+        var_form_Rinv = sum([self.covariance[i,j] * trial[i] * test[j]
+                             for i in range(Vh.dim()) for j in range(Vh.dim())])
+
+        #variational form for the square root of the regularization operator
+        var_form_R_sqrt = sum([self.chol_inv.T[i,j] * trial[j] * test[i]
+                               for i in range(Vh.dim()) for j in range(Vh.dim())])
+
[…]
