### 2018-10-19 93eb0d0
Added GaussianRealPrior unit tests
