+++ hippylib/test/test_gaussian_real_prior.py
@@ -1,0 +1,227 @@
+# Copyright (c) 2016-2018, The University of Texas at Austin 
+# & University of California, Merced.
+#
+# All Rights reserved.
+# See file COPYRIGHT for details.
+#
+# This file is part of the hIPPYlib library. For more information and source code
+# availability see https://hippylib.github.io.
+#
+# hIPPYlib is free software; you can redistribute it and/or modify it under the
+# terms of the GNU General Public License (as published by the Free
+# Software Foundation) version 2.0 dated June 1991.
+from __future__ import absolute_import, division, print_function
+
+import unittest 
+import dolfin as dl
+import numpy as np
+import scipy.stats as scistat
+
+import sys
+sys.path.append('../../')
+from hippylib import *
+
+def get_prior_sample(pr):
+
+    noise = dl.Vector()
+    pr.init_vector(noise, "noise")
+    parRandom.normal(1.0, noise)
+    
+    s = dl.Vector()
+    pr.init_vector(s, 0)
+    pr.sample(noise, s, add_mean=True)
+    
+    return s
+
+class TestGaussianRealPrior(unittest.TestCase):
+    """
+    Test suite for prior.GaussianRealPrior
+    """
+
+    def setUp(self):
+
+        self.dim = np.random.randint(1, high=5)
+
+        self.means = np.random.uniform(-10, high=10., size=self.dim)
+
+        self.chol = np.tril(np.random.randint(1, high=10, size=(self.dim,self.dim)))
+
+        self.cov = np.dot(self.chol, self.chol.T)
+
+        self.precision = np.linalg.inv(self.cov)
+
+        mesh = dl.UnitSquareMesh(3, 3)
+
+        if self.dim > 1:
+            self.Rn = dl.VectorFunctionSpace(mesh, "R", 0, dim=self.dim)
+        else:
+            self.Rn = dl.FunctionSpace(mesh, "R", 0)
+
+        self.test_prior = GaussianRealPrior(self.Rn, self.cov)
+        
+        m = dl.Function(self.Rn)
+        m.vector().zero()
+        m.vector().set_local(self.means)
+        self.test_prior.mean.axpy(1., m.vector())
+
+    
+    def tearDown(self):
+
+        self.test_prior = None
+
+    def test_incompatible_cov(self):
+        #Test whether using an incompatible covariance matrix results in
+        #ValueError
+
+        incompatible_cov = np.identity(self.dim + 1)
+
+        with self.assertRaises(ValueError):
+            temp = GaussianRealPrior(self.Rn, incompatible_cov)
+
+    def test_cost_at_mean(self):
+        #Test whether cost at mean is 0 or not
+
+        expected_cost = 0.0
+        actual_cost = self.test_prior.cost(self.test_prior.mean)
+
+        self.assertAlmostEqual(
+                expected_cost,
+                actual_cost,
+                delta=1e-10)
+
+    def test_cost_sample(self):
+        #Test cost at some random vector
+
+        sample = get_prior_sample(self.test_prior)
+
+        diff = sample.get_local() - self.means
+
+        expected_cost = 0.5 * np.inner(diff, np.dot(self.precision, diff))
+
+        actual_cost = self.test_prior.cost(sample)
+
+        self.assertAlmostEqual(
+                    expected_cost, 
+                    actual_cost, 
+                    delta=1e-10)
+
+    def test_sample_mean(self):
+        #Ensure sample mean converges to true mean
+
+        n_samples = 10000
+
+        samples = [get_prior_sample(self.test_prior).get_local() \
+                   for i in range(n_samples)]
+               
+        sample_mean = np.mean(samples, axis=0)
+
+        rel_err = np.linalg.norm(sample_mean - self.means) \
[…]
