### 2018-10-19 8faf641
Modified travis.yml
