### 2018-10-19 fab9540
Fix imports in test suite
