+++ .travis.yml
@@ -32,4 +32,4 @@ script:
     - docker run --rm -v $(pwd):/home/fenics/hippylib -w /home/fenics/hippylib/applications/poisson quay.io/fenicsproject/stable:2016.2.0 "mpirun -n 1 python model_continuous_obs.py"
     - docker run --rm -v $(pwd):/home/fenics/hippylib -w /home/fenics/hippylib/applications/poisson quay.io/fenicsproject/stable:2016.2.0 "mpirun -n 1 python model_subsurf.py"
     - docker run --rm -v $(pwd):/home/fenics/hippylib -w /home/fenics/hippylib/applications/ad_diff quay.io/fenicsproject/stable:2016.2.0 "mpirun -n 1 python model_ad_diff.py"
-    - docker run --rm -v $(pwd):/home/fenics/hippylib -w /home/fenics/hippylib/hippylib quay.io/fenicsproject/stable:2016.2.0 "python3 -m unittest discover -v"
+    - docker run --rm -v $(pwd):/home/fenics/hippylib -w /home/fenics/hippylib/hippylib/test quay.io/fenicsproject/stable:2016.2.0 "python3 test_gaussian_real_prior.py -v"
