### 2018-10-20 9bd776e
Cleaned up variational forms
