### 2018-10-20 927a977
Removed GaussianRealPrior.logpdf
