### 2018-10-20 1f28f69
Removed GaussianRealPrior.logpdf
