### 2018-10-20 c0dc02e
Add missing list while mapping numpy array
