### 2018-10-22 46ba7ed
Fixed errors related to dl.as_matrix in GaussianRealPrior
