### 2018-10-23 #940 @IgorBaratta
Hi @lheagy  and @uvilla ,

There are no other issues that I would like to be addressed.
Everything looks great now.
Regarding the community guidelines, they are well written and complete. Congrats.

--
Ps: Sorry for the delay, I am attending a conference in China and I didn't have access to my email and github.
