### 2018-10-23 #940 @uvilla
@IgorBaratta and @boehmc,
Thank you very much for your time and constructive comments.

@lheagy,
Thank you for serving as editor for my submission. Today I’ll tag the current state of the repository as release 2.1.1, and after that we’ll be ready to “publish” the manuscript.

Thanks again to everyone.
