+++ doc/conf.py
@@ -29,7 +29,7 @@ author = u'Umberto Villa, Noemi Petra, Omar Ghattas'
 # The short X.Y version
 version = u''
 # The full version, including alpha/beta/rc tags
-release = u'2.1.0'
+release = u'2.1.1'
 
 
 # -- General configuration ---------------------------------------------------
