### 2018-10-23 #940 @uvilla
@lheagy ,

The Zenodo DOI for the lastest hippylib release `2.1.1` (which includes Igor's and Christian's comments) is [10.5281/zenodo.1469675](https://doi.org/10.5281/zenodo.1469675).

The *all-versions* DOI is [10.5281/zenodo.596931](https://doi.org/10.5281/zenodo.596931).

Thank you again to you, @IgorBaratta, @boehmc  for reviewing my code. I truly believe that your reviewed improved the code and how users will interact with hIPPYlib.
