+++ hippylib/modeling/timeDependentPDEProblem.py
@@ -1,0 +1,549 @@
+# Copyright (c) 2016-2018, The University of Texas at Austin & University of
+# California, Merced.
+#
+# All Rights reserved.
+# See file COPYRIGHT for details.
+#
+# This file is part of the hIPPYlib library. For more information and source code
+# availability see https://hippylib.github.io.
+#
+# hIPPYlib is free software; you can redistribute it and/or modify it under the
+# terms of the GNU General Public License (as published by the Free
+# Software Foundation) version 2.0 dated June 1991.
+
+from __future__ import absolute_import, division, print_function
+
+import dolfin as dl
+import numpy as np
+from .PDEProblem import PDEProblem
+from .timeDependentVector import TimeDependentVector
+from .timeDependentOperator import TimeDependentOperator
+from .variables import STATE, PARAMETER, ADJOINT
+from ..algorithms.linalg import Transpose 
+from ..utils.vector2function import vector2Function
+from ..utils.checkDolfinVersion import dlversion
+
+
+class _NonlinearProblem(dl.NonlinearProblem):
+    def __init__(self, L, a, bc):
+        dl.NonlinearProblem.__init__(self)
+        self.L = L
+        self.a = a
+        self.bc = bc
+        self.reset_sparsity = True
+
+    def F(self, b, x):
+        dl.assemble(self.L, tensor=b)
+        if self.bc is not None:
+            for cond in self.bc:
+                cond.apply(b, x)
+
+    def J(self, A, x):
+        dl.assemble(self.a, tensor=A)
+        if self.bc is not None:
+            for cond in self.bc:
+                cond.apply(A)
+
+
+class TimeDependentPDEVariationalProblem(PDEProblem):
+    def __init__(self, Vh, varf_handler, conds, t_init, t_final, dt, is_fwd_linear = False):
+        """
+        varf_handler class
+        conds = [u0, fwd_bc, adj_bc] : initial condition, (essential) fwd_bc, (essential) adj_bc
+        When Vh[STATE] is MixedFunctionSpace, bc's are lists of DirichletBC classes 
+        """
+        
+        self.Vh = Vh
+        self.varf = varf_handler
+
+        if isinstance(conds[1], dl.DirichletBC):
+            self.fwd_bc = [conds[1]]
+        else:
+            self.fwd_bc = conds[1]
+
+        if isinstance(conds[2], dl.DirichletBC):
+            self.adj_bc = [conds[2]]
+        else:
+            self.adj_bc = conds[2]
+
+        self.mesh = self.Vh[STATE].mesh()
+        self.init_cond = dl.interpolate(conds[0], self.Vh[STATE])
+        self.t_init = t_init    
+        self.t_final = t_final
+        self.dt = dt
+        self.times = np.arange(self.t_init, self.t_final+.5*self.dt, self.dt)
+
+        self.linearize_x = None
+        self.solverA = None
+        self.solverAadj = None
+        self.solver_fwd_inc = None
+        self.solver_adj_inc = None
+        self.build_operator = False
+
+        self._A_update = True
+        self._Aadj_update = True
+        self._need_Raa_update = True
+        self._need_C_update = True
+        self._need_Wua_update = True
+
+        self.is_fwd_linear = is_fwd_linear      
+        if not is_fwd_linear:
+            self.NewtonSolver= dl.NewtonSolver()
+            self.NewtonSolver.parameters["absolute_tolerance"] = 1e-10
+            self.NewtonSolver.parameters["relative_tolerance"] = 1e-5
+            self.NewtonSolver.parameters["maximum_iterations"] = 25
+
[…]
