### 2018-10-23 #940 @uvilla
@lheagy and @arfon,

I checked the final PDF and everything looks good.
Again, thank you very much for handling my submission.
